feat: Add version column for optimistic locking in current_account.accounts
- Add migration to create version column with default value of 1
- Update CurrentAccountEntity with Version field
- Implement optimistic locking in Repository.Save() using version check
- Update toEntity/toDomain to map the version field
- Enable TestOptimisticLocking test that was previously skipped

Closes #206

+++ internal/current-account/domain/account.go
@@ -42,7 +42,7 @@ type CurrentAccount struct {
 	OverdraftEnabled      bool
 	OverdraftRate         float64
 	BalanceUpdatedAt      time.Time
-	Version               int
+	Version               int64
 	CreatedAt             time.Time
 	UpdatedAt             time.Time
 }
